docs:  deepSort function description
